Teryaqwithllm/api/requirements.txt
fastapi
uvicorn[standard]

psycopg2-binary
SQLAlchemy
python-dotenv
requests
firebase-admin
pyrebase4
google-auth
google-auth-oauthlib
google-auth-httplib2

numpy
pandas
geopy

passlib[bcrypt]
python-jose[cryptography]

aiohttp
httpx

pydantic
email-validator
joblib
scikit-learn

python-multipart
